+++ src/client/clientcrypt.c
@@ -11,10 +11,10 @@
 #include <string.h>
 #include <stdlib.h>
 
-void encrypt_cli(unsigned long k[], unsigned long tmp[])       //<----This function is copied from the referensed book
+void encrypt_cli(unsigned int k[], unsigned int tmp[])       //<----This function is copied from the referensed book
 {
-    unsigned long y = tmp[0],z = tmp[1];
-    unsigned long delta = 0x9e3779b9, sum = 0;
+    unsigned int y = tmp[0],z = tmp[1];
+    unsigned int delta = 0x9e3779b9, sum = 0;
     int n;
     for (n=0;n<32;n++)
     {
@@ -25,19 +25,19 @@ void encrypt_cli(unsigned long k[], unsigned long tmp[])       //<----This funct
     tmp[0]=y; tmp[1]=z;
 }
 
-unsigned long int encrypt_Handler(char* msg)                                //<------ this function is ispiried by the referensed book
+unsigned int encrypt_Handler(char* msg)                                //<------ this function is ispiried by the referensed book
 {
-    unsigned long k[4];
+    unsigned int k[4];
     k[0]=11111111111111;
     k[1]=22222222222222;    //This is a hardcoded key
     k[2]=33333333333333;
     k[3]=44444444444444;
     char tmp[8];
-    unsigned long int numBlock=0,totalBlock=strlen(msg);  //totalblock is the size of the string
-    unsigned long int msgLenght=strlen(msg)+1,pad=msgLenght%8; //we need to pad so the string is devided by 8 without rest
+    unsigned int numBlock=0,totalBlock=strlen(msg);  //totalblock is the size of the string
+    unsigned int msgLenght=strlen(msg)+1,pad=msgLenght%8; //we need to pad so the string is devided by 8 without rest
     msgLenght=msgLenght+8-pad;
     realloc(msg, msgLenght); // need to allocate more space to this string.
-    unsigned long int i;
+    unsigned int i;
     for (i=totalBlock;i<msgLenght;i++)
     {
         msg[i]=' ';     // padding with spaces
@@ -50,7 +50,7 @@ unsigned long int encrypt_Handler(char* msg)                                //<-
         {
             tmp[i]=msg[i+(numBlock*8)];     // move every block to temporary string and encrypt, then move it back to the message string.
         }
-        encrypt_cli(k,(unsigned long *)tmp);
+        encrypt_cli(k,(unsigned int *)tmp);
         for (i=0;i<8;i++)
         {
             msg[i+(numBlock*8)]=tmp[i];
@@ -60,10 +60,10 @@ unsigned long int encrypt_Handler(char* msg)                                //<-
     return totalBlock*8;
 }
 
-void decrypt(unsigned long k[], unsigned long tmp[])           //<--- This function is also copied from the referensed book
+void decrypt(unsigned int k[], unsigned int tmp[])           //<--- This function is also copied from the referensed book
 {
-    unsigned long y = tmp[0],z = tmp[1];
-    unsigned long delta = 0x9e3779b9, sum = delta << 5;
+    unsigned int y = tmp[0],z = tmp[1];
+    unsigned int delta = 0x9e3779b9, sum = delta << 5;
     int n;
     for (n=0;n<32;n++)
     {
@@ -77,7 +77,7 @@ void decrypt(unsigned long k[], unsigned long tmp[])           //<--- This funct
 void decrypt_Handler(char* msg, int size)        // <--- as encrypt_Handler, is this fumction inspired by the referensed book.
 {
     char tmp[8];
-    unsigned long k[4];
+    unsigned int k[4];
     k[0]=11111111111111;     // same key as in encryption_Handler.
     k[1]=22222222222222;
     k[2]=33333333333333;
@@ -91,7 +91,7 @@ void decrypt_Handler(char* msg, int size)        // <--- as encrypt_Handler, is 
         {
             tmp[i]=msg[i+(numBlock*8)];             // move every block to temporary string and decrypt, then move it back to the message string
         }
-        decrypt (k,(unsigned long *)tmp);
+        decrypt (k,(unsigned int*)tmp);
         for (i=0;i<8;i++)
         {
             msg[i+(numBlock*8)]=tmp[i];
